Add safe teleport check
Added a check when you teleport that will not allow you to teleport if the location has blocks or lava inside of it.

Players with the permission "bnb.warps.unsafebypass" can still teleport, but will get a warning saying it's unsafe and normal players won't be able to access it (this is to avoid losing a warp location).